feat: 가맹점 검색 결과 클릭시 카카오 지도 링크로 이동
